pyproject.toml
[project]
name = "s3-agent-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.62",
    "botocore>=1.40.62",
    "flask>=3.1.2",
    "langgraph>=1.0.2",
    "python-dotenv>=1.2.1",
]
